Created three files to deal with the wavelet sharpness metric
wavelet_metric_and_output contains two functions: 1) wavelet_sharpness, which computes the metric and 2) display_wavelet_decomposition_overlay, which helps visualize what the metric is actually detecting.

wavelet_metric_testing contains two methods, one to test the metric on synthetic data, and one to test the data on a cloud image database.

wavelet_visualization_tools contains a set of plotting tools which are useful for visualizing wavelet decompositions, as well as visualizing certain steps in computing the sharpness metric. This file also contains a testing script that can demonstrate what the methods do, if the file is run on its own.